requirements.txt
pillow
admin-tools-zinnia>=0.1
django-admin-tools>=0.5.2
django-blog-zinnia>=0.15.1
django-debug-toolbar==1.3.0
django-tinymce>=1.5.3
zinnia-wysiwyg-tinymce>=1.2.1
django-static-lor
django-admin-cli
django-dbbackup
flickr-pony
django-request
django-super-favicon
django-recaptcha
akismet
